Меняет положение и вложенность чемодана
