+++ source/js/modules/3d/objects/config.js
@@ -186,3 +186,16 @@ export const airplaneConfig = {
   position: {x: 80, y: -75, z: -55},
   rotate: {x: -60, y: 45, z: 0},
 };
+
+export const suitcaseConfig = {
+  name: `suitcase`,
+  type: `gltf`,
+  path: `3d/gltf/suitcase.gltf`,
+  castShadow: true,
+  receiveShadow: true,
+  scale: {x: 1, y: 1.02, z: 1},
+  finalScale: {x: 1, y: 1, z: 1},
+  position: {x: -280, y: 550, z: 800},
+  finalPosition: {x: -280, y: 515, z: 800},
+  rotate: {x: 0, y: -20, z: 0},
+};
