+++ source/js/modules/3d/story/index.js
@@ -1,9 +1,6 @@
 import * as THREE from 'three';
 
-import {tick, easeInOutQuad} from '../../helpers';
-import {cameraRigSettings, getBubblesConfig, getLightsConfig, getTexturesConfig, objectsToAdd} from './config';
-import {getMaterial, setMeshParams} from '../common';
-import {loadModel} from '../load-object-model';
+import {cameraRigSettings, getBubblesConfig, getLightsConfig, getTexturesConfig} from './config';
 import {getSvgObject} from '../svg-loader';
 
 import StartStory from '../start/start';
@@ -13,9 +10,7 @@ import ThirdStory from './stories/third-story';
 import FourthStory from './stories/fourth-story';
 
 import CameraRig from '../camera-rig';
-
-const SUITCASE_SQUASH_ANIMATION_TIME_SEC = 2.0;
-const SUITCASE_POSITION_ANIMATION_TIME_SEC = 1.4;
+import Suitcase from '../objects/suitcase';
 
 const getCameraRigStageState = (index, config) => {
   return {
@@ -267,25 +262,7 @@ export default class Story {
   }
 
   addSuitcase() {
-    const params = objectsToAdd.suitcase;
-    const material = params.color && getMaterial({color: params.color, ...params.materialReflectivity});
-
-    loadModel(params, material, (mesh) => {
-      const outerGroup = new THREE.Group();
-      outerGroup.castShadow = params.castShadow;
-      outerGroup.receiveShadow = params.receiveShadow;
-      const fluctuationGroup = new THREE.Group();
-
-      setMeshParams(fluctuationGroup, {rotate: params.rotate});
-      setMeshParams(outerGroup, {scale: params.scale});
-      setMeshParams(outerGroup, {position: params.position});
-
-      this.suitcase = {root: outerGroup, fluctation: fluctuationGroup, mesh, params: objectsToAdd.suitcase};
-
-      fluctuationGroup.add(mesh);
-      outerGroup.add(fluctuationGroup);
-      this.rig.addSuitcase(outerGroup);
-    });
+    this.suitcase = new Suitcase(this.rig);
   }
 
   endAnimation() {
@@ -321,36 +298,6 @@ export default class Story {
     this.renderer.render(this.scene, this.camera);
   }
 
-  animateSuitcase() {
-    if (this.suitcase) {
-      if (this.startTime < 0) {
-        this.startTime = new THREE.Clock();
-        return;
-      }
-
-      const t = this.startTime.getElapsedTime();
-      const params = this.suitcase.params;
-
-      // Анимируем падение
-      if (t < SUITCASE_POSITION_ANIMATION_TIME_SEC) {
-        const positionY = tick(params.position.y, params.finalPosition.y, 1);
-        const position = [params.finalPosition.x, positionY, params.finalPosition.z];
-
-        this.suitcase.root.position.set(...position);
-      // Анимируем сжатие / растяжение
-      } else if (t > SUITCASE_POSITION_ANIMATION_TIME_SEC && t < SUITCASE_SQUASH_ANIMATION_TIME_SEC) {
-        let scaleX;
-        let scaleY;
-        let scaleZ;
-
-        scaleY = tick(params.scale.y, params.finalScale.y, easeInOutQuad(t));
-        scaleX = scaleZ = 1 / (Math.sqrt(scaleY)) + 0.002;
-
-        this.suitcase.root.scale.set(scaleX, scaleY, scaleZ);
-      }
-    }
-  }
-
   animate() {
     requestAnimationFrame(this.animate);
 
@@ -379,7 +326,7 @@ export default class Story {
         this.startStory.update();
         break;
       case 1:
-        this.animateSuitcase();
+        this.suitcase.animateAppearance();
         this.storyGroup.children[0].update();
         break;
       case 2:
